added config controller
