+++ src/system/MailerModule/composer.json
@@ -23,7 +23,7 @@
             "displayname": "Mailer Module",
             "url": "mailer",
             "capabilities": {
-                "admin": {"route": "zikulamailermodule_admin_config"},
+                "admin": {"route": "zikulamailermodule_config_config"},
                 "hook_subscriber": {"class": "Zikula\\MailerModule\\Container\\HookContainer"}
             },
             "securityschema": {
